feat: clipboard sync, file transfer, chat, system tray, multi-monitor switch + perf
Nouvelles fonctionnalités :
- Clipboard sync bidirectionnel (arboard) avec commandes Tauri
- File transfer via WebRTC (chunks 48KB, max 100MB, dossier Downloads)
- Chat en session avec ChatPanel.vue (messages temps réel)
- System tray : close→minimize, menu (Ouvrir/ID/Quitter), double-click restore
- Multi-monitor : liste d'écrans distants envoyée au viewer, SelectDisplay
  pour switch de moniteur en live via WebRTC

Optimisations performance :
- Downscale vidéo 1920→1280 dans l'encoder (réduit ~60% la taille JPEG)
- Chunks WebRTC 15KB→60KB avec batch send (moins de round-trips)
- Frame skipping dans le Streamer (skip si envoi précédent en cours)
- Throttle mousemove à 60Hz côté viewer
- MouseMove envoyé avant MouseClick pour positionner le curseur

Protocol : nouveaux variants ControlMessage (ClipboardSync, FileTransfer*,
ChatMessage, SelectDisplay, DisplayListResponse, DisplayInfoProto)

Fixes Clippy : Default impls, io::Error::other, is_multiple_of, is_some_and,
next_back, suppression casts inutiles

+++ client/src/screen_capture.rs
@@ -194,8 +194,8 @@ impl ScreenCapturer for XCapCapturer {
                 Display {
                     id: idx as u32,
                     name: format!("Monitor {}", idx),
-                    width: monitor.width() as u32,
-                    height: monitor.height() as u32,
+                    width: monitor.width(),
+                    height: monitor.height(),
                     x: monitor.x(),
                     y: monitor.y(),
                     is_primary: idx == 0,
@@ -223,7 +223,7 @@ impl ScreenCapturer for XCapCapturer {
     fn get_resolution(&self) -> (u32, u32) {
         if let Some(idx) = self.current_monitor {
             if let Some(monitor) = self.monitors.get(idx) {
-                return (monitor.width() as u32, monitor.height() as u32);
+                return (monitor.width(), monitor.height());
             }
         }
         (1920, 1080) // Fallback
